fix: combine sequence/timestamp atomics
Previously there were two atomics that were tracking both the sequence
and the timestamp. This made it hard to reason about the correctness of
the multithreaded environment. This commit combines both the timestamp
and sequence bits into a single atomic, so that updating them accounts
for both the timestamp update as well as reseting the sequence counter.